Change url pattern to match upstream infra config
This way it goes to the public log files url
instead of private jenkins url

If the value is not overriden in e.g. vars.sh, don't set it
so that the default value can be used. Otherwise, it will be set to
'empty string' which causes zuul to use it's own default which goes
to jenkins.